Generate equivocator hashset efficiently
As can be seen in birthday paradox, drawing unique values from a set with a
naive RNG gets increasingly inefficient as more values are drawn. For reference,
creating a hashset with 30 elements from a 100 element set reaches around 60k
rejections in proptest.
With validators sets not significantly larger than the number of equivocators to
be generated from the set, it is more efficient to shuffle the validator set and
take a slice of the desired size.