Relay votes in messages
Messages sent now include the messages a validator has previously received. This
cuts the messages needed to be passed around significantly, but is also
obviously much more costly with respect to message sized (previously fixed). The
expected number of messages to be passed around when messages are always sent to
all validators (see commented line in message_event(...)) follows
`\sum_{i=0}^{n} \frac{n}{i}`, i.e. the coupon collector's problem.

When the messages are not always sent to all other validators, but instead the
addressee list is randomly sampled, the above sum increases by about 50%.

When relaying votes properly, the code slows to a crawl with about 20 levels of
nesting in the justification. This will have to be addressed.