Handle message update of sender separately from others
To be closer to flow within an actual network, the sender no longer "sends"
themself a new message they produce, but updates it directly. While equivalent
in the current testing framework, this is important for later changes, for
instance where recipients of a message perform validation of the estimate
against its justification.