Sort blockchain parsing queue for debugging
Recipients agree with estimate from validator if they process the queue in the
same order. Specifically, it matters which item in the queue is last. Such a
situation can only come about if another thread has reached the genesis block
and the last block in the queue is as such disconnected. The only time we want
to short-circuit the queue with one remaining block is if all other threads have
merged into the tree where the remaining block is now the lowest height parent.