Only terminate blockchain parsing queue if genesis set is empty
If another thread has reached the genesis block already and we only have one
remaining thread that is not connected yet, its weight will not be accounted for
in the fork choice rule. Furthermore, since the queue is unsorted, it is thus
arbitrary which best block from the lms gets processed last, and thus we
additionally get inconsistent estimates between validators, depending on the
order in which they evaluate the queue.
By only terminating the parsing queue if the genesis set is also empty, we avoid
this issue.