Correctly calculate lm delta in tests::generative_tests::add_message(...)
SenderState::my_last_msg() is not automatically updated when the lms of the
state are. Either, we must implement this, or have my_last_msg() return the lm
stored with the sender's id, but this requires the sender's id to be stored in
SenderState. For now, we just retrieve this lm manually (and assume the lm set
is a singleton).