Add all ltst msgs from justif. in generative_tests::add_message(...)
Would previously only add message and the top level of its justification, which
were only the latest messages new to the sender ever since the sender's last
message. However, a receiver may not have received the prior message from the
sender, and if this message contained messages from other validators that said
receiver also did not receive directly, their state's set of latest messages
would not be updated with these.

Now, we simply extract all latest messages (which has to be done for the state
reconstruction of the sender anyway) and update our state with these directly.