Removed message::Trait. Other types depend on E, V
message::Trait's stuff has moved into message::Message as this trait
is never used apart from message::Message and it is not a parameter
of casper. It was mainly used to contain an Estimator and a
ValidatorName in a weird contrived way. Now, instead of having a
message::Trait, Estimator has a ValidatorName and other generic types
have an estimator and a ValidatorName instead of a message.