Moved vote_count.rs to src/
In order to let the unit tests in the src files use VoteCount, it
must be moved to the src/ folder. It was moved to a new
src/tests_common module.
This change seems to have removed the issue with the dead code
thingy from #43 so I have removed the allow(dead_code) attributes.
Close #60 as it is invalidated by this change.