Replaced PhantomData in Block with actual data.
Changed Block's generic type to trait BlockData rather than
ValidatorName. BlockData uses an associated type ValidatorName.
This data is serialisable which makes it hashable. Blocks can now be
differentiated when they are siblings. Added ValidatorNameBlockData for
the tests, which is a block data structure holding the name of the
validator which created that block. The intent is to have a simple data
structure for the tests.