Added ghost according to 4.30.
This implementation of ghost is the paper's litteral 4.30 definition.
This means its return value is a set of blocks, rather than a single
value. To actually implement the estimator function as the Estimator
trait defines it, we need to choose a single block from that set.

Calling `estimate` rather than `ghost` in `get_height_selected_chain`.